project_gestion/requirements.txt
asgiref==3.9.1
attrs==25.3.0
coverage==7.10.6
Django==5.2.6
django-cors-headers==4.8.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
